+++ backend/ik-alcohol-service/src/main/kotlin/com/iksystem/alcohol/IkAlcoholApplication.kt
@@ -6,8 +6,8 @@ import org.springframework.boot.runApplication
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 
 @SpringBootApplication(scanBasePackages = ["com.iksystem.common", "com.iksystem.alcohol"])
-@EntityScan(basePackages = ["com.iksystem.common"])
-@EnableJpaRepositories(basePackages = ["com.iksystem.common"])
+@EntityScan(basePackages = ["com.iksystem.common", "com.iksystem.alcohol"])
+@EnableJpaRepositories(basePackages = ["com.iksystem.common", "com.iksystem.alcohol"])
 class IkAlcoholApplication
 
 fun main(args: Array<String>) {
