+++ backend/ik-alcohol-service/src/main/kotlin/com/iksystem/alcohol/deviation/model/AlcoholCausalAnalysis.kt
@@ -1,0 +1,11 @@
+package com.iksystem.alcohol.deviation.model
+
+enum class AlcoholCausalAnalysis {
+    MANGLENDE_OPPLAERING,
+    RUTINE_IKKE_FULGT,
+    RUTINE_MANGLER,
+    HOYT_TRYKK_STRESS,
+    UNDERBEMANNING,
+    KOMMUNIKASJON,
+    ANNET,
+}
