+++ backend/ik-alcohol-service/src/main/kotlin/com/iksystem/alcohol/deviation/model/AlcoholReportSource.kt
@@ -1,0 +1,7 @@
+package com.iksystem.alcohol.deviation.model
+
+enum class AlcoholReportSource {
+    EGENRAPPORT,
+    SJENKEKONTROLL,
+    POLITIRAPPORT,
+}
