+++ backend/ik-alcohol-service/src/main/kotlin/com/iksystem/alcohol/penaltypoints/model/PenaltyPoint.kt
@@ -1,0 +1,34 @@
+package com.iksystem.alcohol.penaltypoints.model
+
+import com.iksystem.alcohol.deviation.model.AlcoholDeviation
+import com.iksystem.alcohol.deviation.model.AlcoholDeviationType
+import jakarta.persistence.*
+import java.time.Instant
+
+@Entity
+@Table(name = "penalty_points")
+data class PenaltyPoint(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column(name = "organization_id", nullable = false)
+    val organizationId: Long,
+
+    @ManyToOne
+    @JoinColumn(name = "alcohol_deviation_id")
+    val alcoholDeviation: AlcoholDeviation? = null,
+
+    @Column(nullable = false)
+    val points: Int,
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "violation_type", nullable = false)
+    val violationType: AlcoholDeviationType,
+
+    @Column(length = 500)
+    val description: String? = null,
+
+    @Column(name = "created_at", nullable = false, updatable = false)
+    val createdAt: Instant = Instant.now(),
+)
