--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/alcohol/model/AlcoholCausalAnalysis.kt
@@ -1,11 +1,0 @@
-package com.iksystem.common.deviation.alcohol.model
-
-enum class AlcoholCausalAnalysis {
-    MANGLENDE_OPPLAERING,
-    RUTINE_IKKE_FULGT,
-    RUTINE_MANGLER,
-    HOYT_TRYKK_STRESS,
-    UNDERBEMANNING,
-    KOMMUNIKASJON,
-    ANNET,
-}
