--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/alcohol/model/AlcoholDeviationStatus.kt
@@ -1,7 +1,0 @@
-package com.iksystem.common.deviation.alcohol.model
-
-enum class AlcoholDeviationStatus {
-    OPEN,
-    UNDER_TREATMENT,
-    CLOSED,
-}
