--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/alcohol/model/AlcoholReportSource.kt
@@ -1,7 +1,0 @@
-package com.iksystem.common.deviation.alcohol.model
-
-enum class AlcoholReportSource {
-    EGENRAPPORT,
-    SJENKEKONTROLL,
-    POLITIRAPPORT,
-}
