--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/alcohol/repository/AlcoholDeviationRepository.kt
@@ -1,11 +1,0 @@
-package com.iksystem.common.deviation.alcohol.repository
-
-import com.iksystem.common.deviation.alcohol.model.AlcoholDeviation
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface AlcoholDeviationRepository : JpaRepository<AlcoholDeviation, Long> {
-    fun findAllByOrganizationIdOrderByReportedAtDesc(organizationId: Long): List<AlcoholDeviation>
-    fun findByIdAndOrganizationId(id: Long, organizationId: Long): AlcoholDeviation?
-}
