--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/food/model/FoodDeviationStatus.kt
@@ -1,7 +1,0 @@
-package com.iksystem.common.deviation.food.model
-
-enum class FoodDeviationStatus {
-    OPEN,
-    UNDER_TREATMENT,
-    CLOSED,
-}
