--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/food/model/FoodDeviationType.kt
@@ -1,11 +1,0 @@
-package com.iksystem.common.deviation.food.model
-
-enum class FoodDeviationType {
-    TEMPERATUR,
-    RENHOLD,
-    PERSONLIG_HYGIENE,
-    ALLERGEN,
-    SKADEDYR,
-    MOTTAKSKONTROLL,
-    ANNET,
-}
