--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/model/DeviationModule.kt
@@ -1,9 +1,0 @@
-package com.iksystem.common.deviation.model
-
-/**
- * Functional module the deviation belongs to.
- */
-enum class DeviationModule {
-    IK_MAT,
-    IK_ALKOHOL,
-}
