--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/model/DeviationStatus.kt
@@ -1,10 +1,0 @@
-package com.iksystem.common.deviation.model
-
-/**
- * Lifecycle status for a deviation.
- */
-enum class DeviationStatus {
-    OPEN,
-    IN_PROGRESS,
-    RESOLVED,
-}
