--- backend/ik-common/src/main/kotlin/com/iksystem/common/deviation/repository/DeviationRepository.kt
@@ -1,12 +1,0 @@
-package com.iksystem.common.deviation.repository
-
-import com.iksystem.common.deviation.model.Deviation
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface DeviationRepository : JpaRepository<Deviation, Long> {
-    fun findAllByOrganizationIdOrderByReportedAtDesc(organizationId: Long): List<Deviation>
-
-    fun findByIdAndOrganizationId(id: Long, organizationId: Long): Deviation?
-}
