+++ backend/ik-food-service/src/main/kotlin/com/iksystem/food/deviation/model/FoodDeviationStatus.kt
@@ -1,0 +1,7 @@
+package com.iksystem.food.deviation.model
+
+enum class FoodDeviationStatus {
+    OPEN,
+    UNDER_TREATMENT,
+    CLOSED,
+}
