+++ backend/ik-food-service/src/main/kotlin/com/iksystem/food/deviation/model/FoodDeviationType.kt
@@ -1,0 +1,11 @@
+package com.iksystem.food.deviation.model
+
+enum class FoodDeviationType {
+    TEMPERATUR,
+    RENHOLD,
+    PERSONLIG_HYGIENE,
+    ALLERGEN,
+    SKADEDYR,
+    MOTTAKSKONTROLL,
+    ANNET,
+}
