+++ frontend/nginx.conf
@@ -3,6 +3,31 @@ server {
     root /usr/share/nginx/html;
     index index.html;
 
+    # Deviation endpoints are split across backend services
+    location ^~ /api/v1/deviations/alcohol {
+        proxy_pass http://ik-alkohol:8080;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    location ^~ /api/v1/penalty-points {
+        proxy_pass http://ik-alkohol:8080;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    location ^~ /api/v1/deviations/food {
+        proxy_pass http://ik-mat:8080;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
     # Proxy API requests to backend
     location /api/v1/ {
         proxy_pass http://ik-mat:8080/api/v1/;
